requirements.txt
matplotlib==3.3.4
numpy==1.20.1
pandas==1.1.3
python_dateutil==2.8.1
geopandas==0.9.0
mapclassify==2.4.2
fiona==1.8.18
pyproj==3.0.1
shapely==1.7.1